requirements.txt
Django==1.6.0
django-bootstrap3-datetimepicker
Pillow
django-mailer
django-recaptcha
